+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Fix PostCSS printable plugin to allow printing the advanced backgrounds ([#15](https://github.com/marp-team/marpit/pull/15))
+
 ## v0.0.3 - 2018-05-02
 
 * Implement background image resizing with keyword and scale ([#10](https://github.com/marp-team/marpit/pull/10))